backend/requirements.txt
fastapi
uvicorn
google-genai
beautifulsoup4
httpx
python-dotenv
supabase
groq
pypdf
python-multipart
